支付事务完结

+++ shop-pc/shop-memberinfo/src/main/java/com/yc/shopmemberinfo/dao/ShopMemberinfoMapper.java
@@ -3,7 +3,6 @@ package com.yc.shopmemberinfo.dao;
 import com.yc.bean.MemberInfo;
 import com.yc.vo.Signal;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -31,5 +30,5 @@ public interface ShopMemberinfoMapper {
     //分页查询
     List<MemberInfo> findAllAndPage(MemberInfo memberInfo);
 
-    Integer updatePoint(Signal signal, MemberInfo memberInfo, BigDecimal point);
+    Integer updatePoint(Signal signal, MemberInfo memberInfo, Integer point);
 }
